+++ src/models/budget.py
@@ -1,6 +1,7 @@
 import re
 
 from models.base import Base
+from models.budget_group import BudgetGroup
 from models.taggable import Taggable
 
 class Budget(Taggable, Base):
@@ -57,6 +58,26 @@ class Budget(Taggable, Base):
         data.update(super()._serialize())
 
         return data
+
+
+    @property
+    def monthly_avg(self):
+        """ The average amount spent on this Budget Item per Month """
+        return round(self.amount / self.frequency, 2)
+
+
+    def predict_spending(self, month):
+        """
+        Calculate the amount that should have been spent on this Budget item
+        by the given `month`
+        """
+        amount = 0.0
+        due_months = self.due_months()
+        for due_mnth in due_months:
+            if month >= due_mnth:
+                amount += self.amount
+
+        return round(amount, 2)
 
 
     def frequency_desc(self):
@@ -170,6 +191,7 @@ class Budget(Taggable, Base):
         return due_months
 
 
+    # TODO: Get rid of this in favor of Budget.group()
     @classmethod
     def collate_by_category(self, budget:list):
         """
@@ -193,6 +215,30 @@ class Budget(Taggable, Base):
                 }
             else:
                 budget_map[item.category]["amount"] += item.amount
+
+        return budget_map
+
+
+    @classmethod
+    def group(self, budget:list):
+        """
+        Given a list of Budget items group them by their category.
+
+        Args:
+            budget (list): List of Budget itens
+
+        Return:
+            dict: Mapping of Budget items by their category to BudgetGroups.
+        """
+        budget_map = {}
+        for item in budget:
+            if item.category not in budget_map:
+                budget_group = BudgetGroup(item.category)
+                budget_group.add(item)
+                budget_map[item.category] = budget_group
+            else:
+                budget_group = budget_map[item.category]
+                budget_group.add(item)
 
         return budget_map